elasticsearch/requirements.txt
Faker==25.0.0
numpy==1.24.3
python-dotenv==1.0.1
elasticsearch==8.13.0
